X-Code Mac f21128263a510149fcdfaa30bbed559f
e9254d7ed8d0cfb0027f1e2564b870dded4ab05868ef67cdc39c489f11091537

+++ Code Template/Code_Template/solution.cpp
@@ -218,7 +218,7 @@ void solution(int argc, char* argv[], char* envp[])
     std::cout << "unsigned long : " <<  sizeof(unsigned long) << "B" <<  std::endl;
     std::cout << "unsigned int 64 : " <<  sizeof(uint64_t) << "B" <<  std::endl;
     std::cout << "long double : " <<  sizeof(long double) << "B" <<  std::endl;
-    long double huge_integer = (1ULL << 63);
+    long long huge_integer = (1ULL << 63);
     std::cout << "long double : " << std::dec << huge_integer << std::endl;
 }
